refactor: isolate session ui and reuse ci builds
