+++ package.json
@@ -20,6 +20,7 @@
     "check:loot-parity": "vitest run tests/unit/session-generation-reward-oracles.test.ts tests/unit/session-generation-reward-pipeline-stages.test.ts tests/unit/session-generation-reward-stages.test.ts tests/unit/session-generation-money-boundaries.test.ts tests/unit/session-generation-packing-policy.test.ts tests/unit/session-generation-reward-properties.test.ts tests/unit/session-generation-loot-engine.test.ts tests/integration/generated-run-store.test.ts tests/integration/loot-vertical-slice.test.ts --maxWorkers=2",
     "check:world-planner": "corepack pnpm typecheck && vitest run tests/unit/world-npc-contract.test.ts tests/unit/npc-catalog.test.tsx tests/unit/reference-change-coordinator.test.ts tests/integration/world-npc.test.ts tests/integration/reference-service.test.ts tests/unit/persistence-preflight.test.ts --maxWorkers=2 && corepack pnpm check:version-truth",
     "check:campaign-import": "tsx scripts/run-focused-check.ts campaign-import && corepack pnpm check:version-truth",
+    "check:session-ui": "tsx scripts/check-session-ui.ts",
     "check:version-truth": "tsx scripts/check-version-truth.ts",
     "check:planner-loot": "corepack pnpm typecheck && vitest run tests/unit/architecture-boundaries.test.ts tests/unit/refactor-requirements-alignment.test.ts tests/unit/canonical-fingerprint.test.ts tests/unit/session-generation-rule-provenance.test.ts tests/unit/session-generation-reward-units.test.ts tests/unit/session-generation-reward-pipeline-stages.test.ts tests/unit/session-generation-reward-stages.test.ts tests/unit/session-generation-loot-engine.test.ts tests/unit/session-generation-rational.test.ts tests/unit/session-generation-entropy-streams.test.ts tests/unit/session-generation-run-origin.test.ts tests/unit/generated-encounter-import.test.ts tests/unit/session-planner-draft.test.ts tests/unit/session-planner-scene-inspector.test.tsx tests/unit/session-preparation-contract.test.ts tests/unit/treasure-aggregate-diff.test.ts tests/unit/treasure-draft-reducer.test.ts tests/unit/treasure-draft-fields.test.tsx tests/unit/loot-scene-controller.test.tsx tests/unit/loot-ui.test.tsx tests/unit/money-presenter.test.ts tests/unit/loot-catalog-service.test.ts tests/unit/group-loot-draft.test.ts tests/unit/group-manager-state.test.ts tests/integration/generated-run-store.test.ts tests/integration/generated-encounter-plans.test.ts tests/integration/session-planner-vertical-slice.test.ts tests/integration/loot-projection-store.test.ts tests/integration/loot-vertical-slice.test.ts tests/integration/live-play.test.ts && corepack pnpm build && corepack pnpm test:bundle-budget",
     "profile:encounter-selector": "tsx scripts/profile-encounter-selector.ts",
@@ -37,6 +38,7 @@
     "test:unit": "vitest run tests/unit --exclude tests/unit/architecture-boundaries.test.ts --exclude tests/unit/refactor-requirements-alignment.test.ts --exclude tests/unit/developer-feedback.test.ts --exclude tests/unit/e2e-suite-registry.test.ts --maxWorkers=2",
     "test:unit:portable": "vitest run tests/unit --exclude tests/unit/architecture-boundaries.test.ts --exclude tests/unit/refactor-requirements-alignment.test.ts --exclude tests/unit/developer-feedback.test.ts --exclude tests/unit/e2e-suite-registry.test.ts --exclude tests/unit/local-app-installation.test.ts --exclude tests/unit/local-profile-lock.test.ts --maxWorkers=2",
     "test:unit:linux": "tsx scripts/require-platform.ts linux && vitest run tests/unit/local-app-installation.test.ts tests/unit/local-profile-lock.test.ts --maxWorkers=2",
+    "test:native:persistence": "vitest run tests/integration/campaign-store.test.ts tests/integration/campaign-replacement-safety.test.ts tests/integration/campaign-import.test.ts tests/integration/generated-run-store.test.ts tests/integration/world-npc.test.ts --maxWorkers=2",
     "test:integration": "vitest run tests/integration --maxWorkers=2",
     "test": "corepack pnpm test:architecture && corepack pnpm test:unit && corepack pnpm test:integration",
     "test:smoke": "corepack pnpm build && corepack pnpm test:smoke:built",
@@ -50,7 +52,7 @@
     "test:packaged-qualification-smoke": "corepack pnpm package:qualification && tsx scripts/packaged-smoke.ts --channel release --qualification",
     "test:e2e": "corepack pnpm build && corepack pnpm test:e2e:built",
     "test:e2e:built": "tsx scripts/run-e2e-suites.ts",
-    "test:create:built": "wdio run wdio.conf.ts --suite create",
+    "test:create:built": "tsx scripts/run-e2e-suites.ts --suite campaignCreate --suite campaignHexMap --suite campaignCombat --suite campaignPseudoLocale",
     "test:hex-location:built": "wdio run wdio.conf.ts --suite hexLocation",
     "test:dialogs:built": "wdio run wdio.conf.ts --suite dialogs",
     "test:workspaces:built": "wdio run wdio.conf.ts --suite workspaces",
@@ -58,6 +60,7 @@
     "test:visual:verify": "corepack pnpm build && corepack pnpm test:visual:built",
     "test:visual:update": "tsx scripts/update-visual-golden.ts",
     "test:passive-e2e": "corepack pnpm build && wdio run wdio.passive.conf.ts",
+    "test:passive-e2e:built": "wdio run wdio.passive.conf.ts",
     "qualify:render": "tsx scripts/run-render-qualification.ts",
     "qualify:rp-h": "tsx scripts/qualify-rp-h.ts",
     "qualify:render:validate": "tsx scripts/validate-render-qualification.ts",
@@ -68,6 +71,7 @@
     "check:portable:app": "corepack pnpm build && corepack pnpm test:smoke:built && corepack pnpm test:bundle-budget",
     "check:portable": "corepack pnpm check:portable:fast && corepack pnpm check:portable:app",
     "check:linux": "corepack pnpm test:unit:linux",
+    "check:native": "corepack pnpm test:native:persistence && corepack pnpm build && corepack pnpm test:smoke:built",
     "check:fast": "corepack pnpm format && corepack pnpm lint && corepack pnpm typecheck && corepack pnpm test && corepack pnpm check:references && corepack pnpm check:session-generation && corepack pnpm check:version-truth && corepack pnpm qualify:render:artifacts:check",
     "check:app": "corepack pnpm build && corepack pnpm test:smoke:built && corepack pnpm test:bundle-budget",
     "check:e2e": "corepack pnpm build && corepack pnpm test:e2e:built && corepack pnpm test:visual:built",
@@ -84,6 +88,7 @@
     "delivery:verify-candidate": "tsx scripts/verify-candidate.ts",
     "delivery:promote": "tsx scripts/promote-candidate.ts",
     "delivery:verify-evidence": "tsx scripts/verify-evidence-commit.ts",
+    "delivery:verify-post-promotion": "tsx scripts/verify-post-promotion.ts",
     "package:qualification": "corepack pnpm build:qualification && tsx scripts/package-app.ts --channel release"
   },
   "dependencies": {
