+++ src/renderer/features/session/session-dialogs.css
@@ -9,6 +9,17 @@
   padding: 1rem;
 }
 
+.group-dialog header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 0.6rem;
+}
+
+.group-dialog h2 {
+  margin: 0;
+}
+
 .scene-party-dialog {
   display: grid;
   width: min(34rem, 100%);
@@ -83,8 +94,35 @@
 }
 
 .monster-picker {
+  display: grid;
   max-height: 20rem;
+  gap: 0.45rem;
   overflow: auto;
+  margin: 0;
+  padding: 0;
+  list-style: none;
+}
+
+.monster-picker li {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 0.55rem;
+  border: 1px solid var(--border-1);
+  border-radius: var(--radius-0);
+  background: var(--surface-control);
+  padding: 0.5rem;
+}
+
+.monster-picker span {
+  display: grid;
+  min-width: 0;
+}
+
+.monster-picker small {
+  overflow: hidden;
+  color: var(--text-3);
+  text-overflow: ellipsis;
 }
 
 .monster-picker label {
