+++ src/renderer/features/session/session-groups-panel.css
@@ -6,6 +6,30 @@
     padding: var(--pane-pad);
   }
 
+  .groups-heading {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: var(--space-6);
+    margin-bottom: var(--space-5);
+    border-bottom: var(--border-width-structural) solid
+      var(--accent-rubric-line);
+    padding-bottom: var(--space-4);
+  }
+
+  .groups-heading h2 {
+    margin: 0;
+    font-family: var(--font-smallcaps);
+    font-size: 0.95rem;
+    font-weight: var(--weight-display);
+    letter-spacing: var(--tracking-caps);
+  }
+
+  .session-groups .inline-error,
+  .session-groups .unavailable {
+    color: var(--text-error);
+  }
+
   .group-register {
     display: grid;
     width: 100%;
