+++ tests/e2e/campaign-combat.e2e.ts
@@ -1,0 +1,8 @@
+import { runCampaignCombatScenario } from './support/campaign-walking-scenarios.js'
+
+describe('campaign session combat', () => {
+  it(
+    'builds a scene party, browses monsters and starts a scene group combat',
+    runCampaignCombatScenario
+  )
+})
