+++ tests/e2e/campaign-create.e2e.ts
@@ -1,0 +1,5 @@
+import { runCampaignCreationScenario } from './support/campaign-walking-scenarios.js'
+
+describe('campaign creation and switching', () => {
+  it('creates and switches the selected campaign', runCampaignCreationScenario)
+})
