+++ tests/e2e/campaign-hex-map.e2e.ts
@@ -1,0 +1,8 @@
+import { runCampaignHexMapScenario } from './support/campaign-walking-scenarios.js'
+
+describe('campaign hex map', () => {
+  it(
+    'keeps a newly created hex map inside the workspace',
+    runCampaignHexMapScenario
+  )
+})
