+++ tests/e2e/campaign-pseudo-locale.e2e.ts
@@ -1,0 +1,8 @@
+import { runCampaignPseudoLocaleScenario } from './support/campaign-walking-scenarios.js'
+
+describe('campaign pseudo locale', () => {
+  it(
+    'survives the pseudo locale without accessibility regressions',
+    runCampaignPseudoLocaleScenario
+  )
+})
