+++ tests/unit/ci-workflow-policy.test.ts
@@ -7,11 +7,15 @@ const packageJson = JSON.parse(readFileSync('package.json', 'utf8')) as {
 }
 
 describe('CI platform partitions', () => {
-  it('runs portable checks on every supported host', () => {
+  it('runs portable checks once and keeps native SQLite/runtime checks per OS', () => {
+    expect(workflow.match(/pnpm check:portable:fast/g)).toHaveLength(1)
+    expect(workflow).toContain('name: Native · ${{ matrix.os }}')
+    expect(workflow).toContain('- os: windows-2022')
+    expect(workflow).toContain('- os: macos-latest')
+    expect(workflow).toContain('run: pnpm check:native')
     expect(workflow).toContain(
-      'matrix: { os: [ubuntu-latest, windows-2022, macos-latest] }'
+      "native-${{ runner.os }}-${{ runner.arch }}-node22-pnpm10-${{ hashFiles('pnpm-lock.yaml') }}"
     )
-    expect(workflow).toContain('run: pnpm check:portable')
     expect(packageJson.scripts['test:unit:portable']).toContain(
       '--exclude tests/unit/local-app-installation.test.ts'
     )
@@ -31,15 +35,37 @@ describe('CI platform partitions', () => {
       'tests/unit/local-profile-lock.test.ts'
     )
     const linuxJob = workflow.slice(
-      workflow.indexOf('  linux-runtime:'),
-      workflow.indexOf('  e2e:')
+      workflow.indexOf('  linux-build:'),
+      workflow.indexOf('  linux-package:')
     )
     expect(linuxJob).toContain('runs-on: ubuntu-latest')
-    expect(linuxJob).toContain('pnpm test:packaged-smoke')
-    expect(linuxJob).toContain('pnpm test:packaged-qualification-smoke')
+    expect(linuxJob).toContain('pnpm check:linux')
+    const packageJob = workflow.slice(
+      workflow.indexOf('  linux-package:'),
+      workflow.indexOf('  e2e:')
+    )
+    expect(packageJob).toContain('pnpm package:development:built')
+    expect(packageJob).toContain('pnpm test:packaged-smoke:built')
+    expect(workflow).toContain('name: Linux qualification · packaged harness')
+    expect(workflow).toContain('pnpm test:packaged-qualification-smoke')
   })
 
-  it('shards isolated E2E fixtures and retains complete failure evidence', () => {
+  it('builds Linux once and validates the same receipt in every consumer', () => {
+    const linuxBuild = workflow.slice(
+      workflow.indexOf('  linux-build:'),
+      workflow.indexOf('  linux-package:')
+    )
+    expect(linuxBuild.match(/^\s+- run: pnpm build$/gm)).toHaveLength(1)
+    const consumers = workflow.slice(workflow.indexOf('  linux-package:'))
+    expect(consumers).not.toMatch(/^\s+- run: pnpm build$/m)
+    expect(workflow).toContain('name: linux-app-${{ github.sha }}')
+    expect(workflow.match(/actions\/download-artifact@v4/g)).toHaveLength(4)
+    expect(
+      workflow.match(/assert-built-workspace\.ts --channel development/g)
+    ).toHaveLength(5)
+  })
+
+  it('shards isolated functional and visual E2E while retaining evidence', () => {
     for (const path of [
       '.tmp/e2e-runs',
       '.tmp/visual-diffs',
@@ -47,6 +73,16 @@ describe('CI platform partitions', () => {
     ])
       expect(workflow).toContain(path)
     expect(workflow).toContain('if-no-files-found: error')
-    expect(workflow).toContain('--suite workspaces --suite create')
+    expect(workflow).toContain('--suite workspaces --suite campaignCreate')
+    expect(workflow).toContain('name: Linux Visual · goldens')
+    expect(workflow).toContain('pnpm test:visual:built')
+  })
+
+  it('runs only candidate attestation after the exact SHA reaches main', () => {
+    const postPromotion = workflow.slice(workflow.indexOf('  post-promotion:'))
+    expect(postPromotion).toContain("if: github.event_name == 'push'")
+    expect(postPromotion).toContain('pnpm delivery:verify-post-promotion')
+    expect(postPromotion).not.toContain('pnpm build')
+    expect(postPromotion).not.toContain('pnpm test')
   })
 })
